restructured constants
