+++ components/StructuralElements/footer.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 import { LINKS } from "./header";
 import { ConnectButton } from "../ControlElements/connect-button";
 import { useIsMobile } from "../../lib/hooks/useWindowSize";
-import { FOOTER_LOGO } from "../../lib/constants";
+import { FOOTER_LOGO } from "../../lib/constants/url";
 
 const Container = styled.div`
   bottom: 0;
