+++ lib/constants/abi.ts
@@ -1,0 +1,41 @@
+export const ERC20_ABI = [
+  // Read-Only Functions
+  "function name() public view returns (string)",
+  "function symbol() public view returns (string)",
+  "function decimals() public view returns (uint8)",
+  "function balanceOf(address _owner) public view returns (uint256 balance)",
+  "function totalSupply() public view returns (uint256)",
+];
+
+export const ERC20_ABI_MAKER = [
+  "function name() public view returns (bytes32)",
+  "function symbol() public view returns (bytes32)",
+  "function decimals() public view returns (uint8)",
+  "function balanceOf(address _owner) public view returns (uint256 balance)",
+  "function totalSupply() public view returns (uint256)",
+];
+
+// for ethers-multicall
+export const ERC20JsonAbi = [
+  {
+    constant: true,
+    inputs: [
+      {
+        internalType: "address",
+        name: "account",
+        type: "address",
+      },
+    ],
+    name: "balanceOf",
+    outputs: [
+      {
+        internalType: "uint256",
+        name: "",
+        type: "uint256",
+      },
+    ],
+    payable: false,
+    stateMutability: "view",
+    type: "function",
+  },
+];
