+++ lib/constants/env.ts
@@ -1,0 +1,4 @@
+export const ETH_BLOCK_HEIGHT_PADDING = 10;
+export const GOERLI_MULTICALL = "0x77dca2c955b15e9de4dbbcf1246b4b85b651e50e";
+export const GOERLI_CHAINID = 5;
+export const EMPTY_ADDRESS = "0x0000000000000000000000000000000000000000";
