+++ lib/constants/url.ts
@@ -1,50 +1,3 @@
-export const ETH_BLOCK_HEIGHT_PADDING = 10;
-export const GOERLI_MULTICALL = "0x77dca2c955b15e9de4dbbcf1246b4b85b651e50e";
-export const GOERLI_CHAINID = 5;
-export const EMPTY_ADDRESS = "0x0000000000000000000000000000000000000000";
-
-export const ERC20_ABI = [
-  // Read-Only Functions
-  "function name() public view returns (string)",
-  "function symbol() public view returns (string)",
-  "function decimals() public view returns (uint8)",
-  "function balanceOf(address _owner) public view returns (uint256 balance)",
-  "function totalSupply() public view returns (uint256)",
-];
-
-export const ERC20_ABI_MAKER = [
-  "function name() public view returns (bytes32)",
-  "function symbol() public view returns (bytes32)",
-  "function decimals() public view returns (uint8)",
-  "function balanceOf(address _owner) public view returns (uint256 balance)",
-  "function totalSupply() public view returns (uint256)",
-];
-
-// for ethers-multicall
-export const ERC20JsonAbi = [
-  {
-    constant: true,
-    inputs: [
-      {
-        internalType: "address",
-        name: "account",
-        type: "address",
-      },
-    ],
-    name: "balanceOf",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-];
-
 //LOGOS
 export const FALLBACK_TOKEN_ICON = "/media/eth-logo.png";
 export const HEADER_LOGO = "/media/logo_voice.svg";
