+++ lib/hooks/process/index.ts
@@ -1,3 +1,4 @@
+export { useCensusProof } from "./useCensusProof";
 export { useProcessDates } from "./useProcessDates";
 export { useProcessResults } from "./useProcessResults";
 export { useVote } from "./useVote";
